More on line charge functionality
Fixed a bug where mouseUp would stop the draw method prematurely if you
moved the element and let go of the mouse too quickly, causing the
element to jump positions the next time draw method is called.

Added the updateChargeListLineCharge() method which adds a list element
customized for line charges to the elements list.

Added findAngle() method to find the angle of a line charge.

Using the findAngle() method, implemented outlining of a line charge
element when it is selected.